+++ HARK/ConsumptionSaving/tests/test_IndShockConsumerType.py
@@ -6,7 +6,7 @@ from HARK.ConsumptionSaving.ConsIndShockModel import (
 )
 import numpy as np
 import unittest
-from copy import copy
+from copy import copy, deepcopy
 
 
 class testIndShockConsumerType(unittest.TestCase):
@@ -24,8 +24,8 @@ class testIndShockConsumerType(unittest.TestCase):
         self.agent.get_shocks()
 
         self.assertEqual(self.agent.shocks['PermShk'][0], 1.0427376294215103)
-        self.assertEqual(self.agent.shocks['PermShk'][1], 0.9278094171517413)
-        self.assertEqual(self.agent.shocks['TranShk'][0], 0.881761797501595)
+        self.assertAlmostEqual(self.agent.shocks['PermShk'][1], 0.9278094171517413)
+        self.assertAlmostEqual(self.agent.shocks['TranShk'][0], 0.881761797501595)
 
     def test_ConsIndShockSolverBasic(self):
         LifecycleExample = IndShockConsumerType(**init_lifecycle)
@@ -140,9 +140,9 @@ class testBufferStock(unittest.TestCase):
         GICRaw_fail_dictionary = dict(self.base_params)
         GICRaw_fail_dictionary["Rfree"] = 1.08
         GICRaw_fail_dictionary["PermGroFac"] = [1.00]
+        GICRaw_fail_dictionary["cycles"] = 0 # cycles=0 makes this an infinite horizon consumer
 
         GICRawFailExample = IndShockConsumerType(
-            cycles=0,  # cycles=0 makes this an infinite horizon consumer
             **GICRaw_fail_dictionary
         )
 
@@ -157,8 +157,8 @@ class testBufferStock(unittest.TestCase):
         self.assertFalse(GICRawFailExample.conditions["GICRaw"])
 
     def test_infinite_horizon(self):
-        baseEx_inf = IndShockConsumerType(cycles=0, **self.base_params)
-
+        baseEx_inf = IndShockConsumerType(**self.base_params)
+        baseEx_inf.assign_parameters(cycles = 0)
         baseEx_inf.solve()
         baseEx_inf.unpack("cFunc")
 
@@ -229,7 +229,7 @@ IdiosyncDict = {
 class testIndShockConsumerTypeExample(unittest.TestCase):
     def test_infinite_horizon(self):
         IndShockExample = IndShockConsumerType(**IdiosyncDict)
-        IndShockExample.cycles = 0  # Make this type have an infinite horizon
+        IndShockExample.assign_parameters(cycles = 0)  # Make this type have an infinite horizon
         IndShockExample.solve()
 
         self.assertAlmostEqual(IndShockExample.solution[0].mNrmStE, 1.5488165705077026)
@@ -387,7 +387,8 @@ class testStablePoints(unittest.TestCase):
         # Theory" paper.
 
         # Create and solve the agent
-        baseAgent_Inf = IndShockConsumerType(cycles=0, verbose=0, **bst_params)
+        baseAgent_Inf = IndShockConsumerType(verbose=0, **bst_params)
+        baseAgent_Inf.assign_parameters(cycles = 0)
         baseAgent_Inf.solve()
 
         # Extract stable points
@@ -398,3 +399,181 @@ class testStablePoints(unittest.TestCase):
         decimalPlacesTo = 10
         self.assertAlmostEqual(mNrmStE, 1.37731133865, decimalPlacesTo)
         self.assertAlmostEqual(mNrmTrg, 1.39101653806, decimalPlacesTo)
+
+
+JACDict={
+    # Parameters shared with the perfect foresight model
+    "CRRA":2,                             # Coefficient of relative risk aversion
+    "Rfree": 1.05**.25,                  # Interest factor on assets
+    "DiscFac": 0.972,                    # Intertemporal discount factor
+    "LivPrb" : [.99375],                    # Survival probability
+    "PermGroFac" :[1.00],                 # Permanent income growth factor
+
+    # Parameters that specify the income distribution over the lifecycle
+   
+    "PermShkStd" :  [(0.01*4/11)**0.5],    # Standard deviation of log permanent shocks to income
+    "PermShkCount" : 5,                    # Number of points in discrete approximation to permanent income shocks
+    "TranShkStd" : [.2],                   # Standard deviation of log transitory shocks to income
+    "TranShkCount" : 5,                    # Number of points in discrete approximation to transitory income shocks
+    "UnempPrb" : 0.05,                     # Probability of unemployment while working
+    "IncUnemp" : 0.1,      # Unemployment benefits replacement rate
+    "UnempPrbRet" : 0.0005,                # Probability of "unemployment" while retired
+    "IncUnempRet" : 0.0,                   # "Unemployment" benefits when retired
+    "T_retire" : 0,                        # Period of retirement (0 --> no retirement)
+    "tax_rate" : 0.2,      # Flat income tax rate (legacy parameter, will be removed in future)
+
+    # Parameters for constructing the "assets above minimum" grid
+    "aXtraMin" : 0.001,                    # Minimum end-of-period "assets above minimum" value
+    "aXtraMax" : 15,                       # Maximum end-of-period "assets above minimum" value
+    "aXtraCount" : 48,                     # Number of points in the base grid of "assets above minimum"
+    "aXtraNestFac" : 3,                    # Exponential nesting factor when constructing "assets above minimum" grid
+    "aXtraExtra" : [None],                 # Additional values to add to aXtraGrid
+
+    # A few other parameters
+    "BoroCnstArt" : 0.0,                   # Artificial borrowing constraint; imposed minimum level of end-of period assets
+    "vFuncBool" : True,                    # Whether to calculate the value function during solution
+    "CubicBool" : False,                   # Preference shocks currently only compatible with linear cFunc
+    "T_cycle" : 1,                         # Number of periods in the cycle for this agent type
+
+    # Parameters only used in simulation
+    "AgentCount" : 5000,                 # Number of agents of this type
+    "T_sim" : 100,                         # Number of periods to simulate
+    "aNrmInitMean" : np.log(2)-(.5**2)/2,# Mean of log initial assets
+    "aNrmInitStd"  : .5,                   # Standard deviation of log initial assets
+    "pLvlInitMean" : 0,                    # Mean of log initial permanent income
+    "pLvlInitStd"  : 0,                    # Standard deviation of log initial permanent income
+    "PermGroFacAgg" : 1.0,                 # Aggregate permanent income growth factor
+    "T_age" : None,                        # Age after which simulated agents are automatically killed
+  
+}
+
+
+class testPerfMITShk(unittest.TestCase):
+    
+    def jacobian(self):
+        
+        class Test_agent(IndShockConsumerType):
+            
+            def __init__(self, cycles= 0, **kwds):
+                
+                IndShockConsumerType.__init__(self, cycles = 0, **kwds)
+            
+            def get_Rfree(self):
+                """
+                Returns an array of size self.AgentCount with self.Rfree in every entry.
+                Parameters
+                ----------
+                None
+                Returns
+                -------
+                RfreeNow : np.array
+                     Array of size self.AgentCount with risk free interest rate for each agent.
+                """
+                
+                if type(self.Rfree) == list:
+                    RfreeNow = self.Rfree[self.t_sim]* np.ones(self.AgentCount)
+                else:
+                    RfreeNow = ss.Rfree * np.ones(self.AgentCount)
+                    
+                return RfreeNow
+    
+        ss = Test_agent(**JACDict )
+        ss.cycles = 0
+        ss.T_sim= 1200
+        ss.solve()
+        ss.initialize_sim()
+        ss.simulate()
+        
+        class Test_agent2(Test_agent):
+            
+             def transition(self):
+                
+                pLvlPrev = self.state_prev['pLvl']
+                aNrmPrev = self.state_prev['aNrm']
+                RfreeNow = self.get_Rfree()
+            
+                # Calculate new states: normalized market resources and permanent income level
+                pLvlNow = pLvlPrev*self.shocks['PermShk']  # Updated permanent income level
+                # Updated aggregate permanent productivity level
+                PlvlAggNow = self.state_prev['PlvlAgg']*self.PermShkAggNow
+                # "Effective" interest factor on normalized assets
+                ReffNow = RfreeNow/self.shocks['PermShk']
+                bNrmNow = ReffNow*aNrmPrev         # Bank balances before labor income
+                mNrmNow = bNrmNow + self.shocks['TranShk']  # Market resources after income
+                
+            
+                if self.t_sim == 0:
+                        
+                        mNrmNow = ss.state_now['mNrm']
+                        pLvlNow = ss.state_now['pLvl']
+            
+                return pLvlNow, PlvlAggNow, bNrmNow, mNrmNow, None        
+        
+        
+    
+        listA_g = []
+        params = deepcopy(JACDict)
+        params['T_cycle']= 200
+        params['LivPrb']= params['T_cycle']*[ss.LivPrb[0]]
+        params['PermGroFac']=params['T_cycle']*[1]
+        params['PermShkStd'] = params['T_cycle']*[(0.01*4/11)**0.5]
+        params['TranShkStd']= params['T_cycle']*[.2]
+        params['Rfree'] = params['T_cycle']*[ss.Rfree]
+        
+        ss_dx = Test_agent2(**params )
+        ss_dx.pseudo_terminal = False
+        ss_dx.PerfMITShk = True
+        ss_dx.track_vars = ['aNrm','mNrm','cNrm','pLvl','aLvl']
+        ss_dx.cFunc_terminal_ = deepcopy(ss.solution[0].cFunc)
+        ss_dx.T_sim = params['T_cycle']
+        ss_dx.cycles= 1
+        ss_dx.IncShkDstn = params['T_cycle']*ss_dx.IncShkDstn
+        ss_dx.del_from_time_inv('Rfree')
+        ss_dx.add_to_time_vary('Rfree')
+        
+        ss_dx.solve()
+        ss_dx.initialize_sim()
+        ss_dx.simulate()
+        
+        
+        for j in range(ss_dx.T_sim):
+        
+            Ag = np.mean(ss_dx.history['aLvl'][j,:])
+            listA_g.append(Ag)
+        
+        A_dx0 = np.array(listA_g)
+        
+        
+        ##############################################################################
+        
+        example = Test_agent2(**params )
+        example.pseudo_terminal=False 
+        example.cFunc_terminal_ = deepcopy(ss.solution[0].cFunc)
+        example.T_sim = params['T_cycle']
+        example.cycles = 1
+        example.PerfMITShk = True
+        example.track_vars = ['aNrm','mNrm','cNrm','pLvl','aLvl']
+        example.del_from_time_inv('Rfree')
+        example.add_to_time_vary('Rfree')
+        example.IncShkDstn = params['T_cycle']*example.IncShkDstn
+        
+        AHist =[]
+        listA = []
+        dx = .001
+        i = 50
+        
+        example.Rfree = i *[ss.Rfree] + [ss.Rfree + dx] + (params['T_cycle']  - i - 1)*[ss.Rfree]
+         
+        example.solve()
+        example.initialize_sim()
+        example.simulate()
+        
+        for j in range(example.T_sim):
+         
+            a = np.mean(example.history['aLvl'][j,:])
+            listA.append(a)
+            
+        AHist.append(np.array(listA))
+        JACA = (AHist[0]-A_dx0)/(dx)
+        
+        self.assertAlmostEqual(JACA[175], 6.441930322509393e-06)