+++ HARK/core.py
@@ -263,7 +263,6 @@ class AgentType(Model):
     def __init__(
         self,
         solution_terminal=None,
-        cycles=1,
         pseudo_terminal=True,
         tolerance=0.000001,
         seed=0,
@@ -275,7 +274,6 @@ class AgentType(Model):
             solution_terminal = NullFunc()
 
         self.solution_terminal = solution_terminal  # NOQA
-        self.cycles = cycles  # NOQA
         self.pseudo_terminal = pseudo_terminal  # NOQA
         self.solve_one_period = NullFunc()  # NOQA
         self.tolerance = tolerance  # NOQA